feat(notebook): redesign creation flow as 4-step wizard
The /notebooks/new route still carried its old modal logic — a 2-dot
stepper, click-to-edit name/description with autoFocus-on-click, the
Wolke source buried inside step 2, and the public/ownership controls
jammed at the bottom of one long form. With creation now on a full-
width page, replace it with a guided 4-step wizard (Quellen → Details
→ Sichtbarkeit → Überprüfen) using the canonical MultiStepForm
primitive from @gruenerator/ui (matching WolkeSetupWizard). Wolke is
now a first-class source on step 1, name/description are proper
labelled fields, ownership becomes clickable cards, and a final
summary card replaces the "did I miss anything?" submit anxiety.

Also splits the 1000-line monolith into a NotebookEditor/ folder for
maintainability: index.tsx orchestrates create vs edit, a single
useNotebookEditorState hook owns all shared state and callbacks,
four step components consume a typed ReturnType<typeof ...> bundle,
and DocumentCard + LabelsField are shared between wizard and edit
form. Edit mode is untouched apart from migrating the document grid's
empty state to <Empty> for design-system consistency.
